♻️ Extract common combobox logic out of MetadataCombobox
